feat: バックグラウンド（放置）時の時間経過ダメージ補償を実装\n\n- アプリ復帰時に経過時間分のダメージを一括加算するロジックを追加\n- syncStepsの戻り値をダメージ量(double)に変更\n- HomeScreenとMainScreenで型変換エラーを修正

+++ lib/main.dart
@@ -46,7 +46,7 @@ class _EggWalkerAppState extends ConsumerState<EggWalkerApp>
     if (expGained > 0 && mounted) {
       // 少し遅延させてからダイアログを表示（画面描画完了待ち）
       Future.delayed(const Duration(milliseconds: 500), () {
-        _showWelcomeBackDialog(expGained);
+        _showWelcomeBackDialog(expGained.toInt());
       });
     }
   }
